Move Range to core
