+++ dnainator-core/src/main/java/nl/tudelft/dnainator/annotation/Annotation.java
@@ -1,5 +1,7 @@
 package nl.tudelft.dnainator.annotation;
 
+import nl.tudelft.dnainator.core.Range;
+
 import java.util.Collection;
 
 /**
