+++ dnainator-core/src/main/java/nl/tudelft/dnainator/annotation/impl/AnnotationCollectionImpl.java
@@ -2,7 +2,7 @@ package nl.tudelft.dnainator.annotation.impl;
 
 import nl.tudelft.dnainator.annotation.Annotation;
 import nl.tudelft.dnainator.annotation.AnnotationCollection;
-import nl.tudelft.dnainator.annotation.Range;
+import nl.tudelft.dnainator.core.Range;
 import nl.tudelft.dnainator.parser.Iterator;
 
 import java.util.Collection;
