+++ dnainator-core/src/main/java/nl/tudelft/dnainator/graph/Graph.java
@@ -2,7 +2,7 @@ package nl.tudelft.dnainator.graph;
 
 import nl.tudelft.dnainator.annotation.Annotation;
 import nl.tudelft.dnainator.annotation.AnnotationCollection;
-import nl.tudelft.dnainator.annotation.Range;
+import nl.tudelft.dnainator.core.Range;
 import nl.tudelft.dnainator.core.EnrichedSequenceNode;
 import nl.tudelft.dnainator.core.impl.Cluster;
 import nl.tudelft.dnainator.core.impl.TreeNode;
