+++ dnainator-core/src/main/java/nl/tudelft/dnainator/graph/impl/nodes/Neo4jAnnotation.java
@@ -1,7 +1,7 @@
 package nl.tudelft.dnainator.graph.impl.nodes;
 
 import nl.tudelft.dnainator.annotation.Annotation;
-import nl.tudelft.dnainator.annotation.Range;
+import nl.tudelft.dnainator.core.Range;
 import nl.tudelft.dnainator.graph.impl.properties.AnnotationProperties;
 import nl.tudelft.dnainator.graph.impl.properties.NodeLabels;
 import nl.tudelft.dnainator.graph.impl.properties.RelTypes;
