+++ dnainator-core/src/test/java/nl/tudelft/dnainator/annotation/DRMutationTest.java
@@ -2,6 +2,7 @@ package nl.tudelft.dnainator.annotation;
 
 import nl.tudelft.dnainator.annotation.impl.DRMutation;
 
+import nl.tudelft.dnainator.core.Range;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
