+++ dnainator-core/src/test/java/nl/tudelft/dnainator/annotation/impl/AnnotationImpl.java
@@ -4,7 +4,7 @@ import java.util.Collection;
 import java.util.Collections;
 
 import nl.tudelft.dnainator.annotation.Annotation;
-import nl.tudelft.dnainator.annotation.Range;
+import nl.tudelft.dnainator.core.Range;
 
 /**
  * A default implementation of an Annotation.
